Stop relying on group split matching to keep separators
"string.path".split(/\./) should return: ["string", ".", "path"]

This is not the case in IE8.
IE returns:  ["string", "path"]

Instead do the split discarding the separators, get model and keypath
from this array. Check unsplit path  to se if the string has a bypass
separator to set the option flag.

This should probably be tested.