each-* will now insert between siblings [Closes #106]
This commit fixes an issue with the each-* binding where all but the
first element would be inserted into incorrect positions when the
binding is used on an element with siblings.

The issue occurred because the code was passing a Rivets.View instance
in to insertBefore for all but the first element, which just caused the
element to be inserted last.